requirements.txt
Django
django-environ
model-import-export
django-import-export
psycopg2
jsonfield
django-rest-framework
structlog
whitenoise
pyngo



# Rest framework

djangorestframework
markdown       
django-filter  
pygments

# quickbooks
python-quickbooks
intuit-oauth

# xero

pyxero

# Admin Themes
django-unfold